final touches

+++ Server/files.c
@@ -18,7 +18,7 @@ int closeFile(int fd, int *err)
 	return res;
 }
 
-int readFile(char* buff, size_t length, int fd, off_t offset, int *err)
+ssize_t readFile(char* buff, size_t length, int fd, off_t offset, int *err)
 {
 	off_t check;
 	ssize_t bytes;
@@ -40,7 +40,7 @@ int readFile(char* buff, size_t length, int fd, off_t offset, int *err)
 	return bytes;
 }
 
-int writeFile(char *buff, size_t length, int fd, off_t offset, int *err){
+ssize_t writeFile(char *buff, size_t length, int fd, off_t offset, int *err){
 	off_t check;
 	ssize_t bytes;
