+++ Server/files.h
@@ -11,7 +11,7 @@
 
 int openFile(char* pathname, int *err);
 int closeFile(int fd, int *err);
-int readFile(char* buff, size_t length, int fd, off_t offset, int *err);
-int writeFile(char *buff, size_t length, int fd, off_t offset, int *err);
+ssize_t readFile(char* buff, size_t length, int fd, off_t offset, int *err);
+ssize_t writeFile(char *buff, size_t length, int fd, off_t offset, int *err);
 
 #endif
